+++ src/const.ts
@@ -1,4 +1,6 @@
-export const APP_VERSION = __APP_VERSION__
+import packageJson from '../package.json' with { type: 'json' }
+
+export const APP_VERSION = packageJson.version
 export const OFFSET_APP_TYPE = 0x10000
 export const OFFSET_DATA_TYPE = 0x1000
 export const PARTITION_TABLE_SIZE = 0x1000 // 4KB partition table size