+++ vite.config.ts
@@ -2,7 +2,6 @@ import { fileURLToPath, URL } from 'node:url'
 
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-import packageJson from './package.json' with { type: 'json' }
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -12,9 +11,6 @@ export default defineConfig({
   plugins: [
     vue(),
   ],
-  define: {
-    __APP_VERSION__: JSON.stringify(packageJson.version)
-  },
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url))